revamp wims api

+++ src/main/java/com/ascentdev/wims/entity/RackEntity.java
@@ -58,7 +58,7 @@ public class RackEntity {
   int modifiedById;
   
   @Column(name="txn_mawb_id")
-  int txnMawbId;
+  long txnMawbId;
   
   @Column(name="txn_hawb_id")
   int txnHawbId;
